refactor(deps): changed folder structure for custom storybook
